LUCENE-4514: make scorer.freq() well defined: number of matches in doc
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/cleanup2878@1403702 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/search/ExactPhraseScorer.java
@@ -184,7 +184,7 @@ final class ExactPhraseScorer extends Scorer {
   }
 
   @Override
-  public float freq() {
+  public int freq() {
     return freq;
   }
